Move file format tests to separate test class + adds json Check tests.
